% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plots.R
\name{wait_plot}
\alias{wait_plot}
\title{plots}
\usage{
wait_plot(wait_frame)
}
\arguments{
\item{wait_frame}{dataframe. You probably made this with a function like
\code{\link{simple_input}}}
}
\value{
A plot showing the size of the wait list over time
}
\description{
Plot the waiting list over time, optionally varying
the input/ output rate at a "specific" break point in the data
}
