% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simple.R
\name{simple_input}
\alias{simple_input}
\title{simple_input}
\usage{
simple_input(
  wait_list,
  rate_in,
  rate_out,
  start_date,
  date_unit,
  length_output
)
}
\arguments{
\item{wait_list}{numeric. Size of wait list at initialisation}

\item{rate_in}{numeric. Rate (expressed in the unit of time selected in
date_unit) of referrals in to assessment}

\item{rate_out}{numeric. Rate (expressed in the unit of time selected in
date_unit) of assessments out}

\item{start_date}{date. Date of initialisation}

\item{date_unit}{string. "day", "week", or "month". Date interval to output
the dataframe in (in the same units as rate_*)}

\item{length_output}{int. Length of output dataframe (i.e. number of rows)}
}
\value{
dataframe with two columns- date and waitlist size
}
\description{
A simple function to model input waiting lists (that is,
ones with only referrals as inputs). Takes as
arguments only initial wait list size and rate of input and output,
start date, date increment (day, week, month), and length of output
}
